requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
Box2D==2.3.10
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.37
docker-pycreds==0.4.0
etils==1.3.0
eval_type_backport==0.2.2
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.16.1
fonttools==4.55.4
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
glfw==2.8.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.69.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
gymnasium-robotics==1.3.1
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.5
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
mujoco==3.1.6
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pettingzoo==1.24.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pygame==2.6.1
PyOpenGL==3.1.9
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sentry-sdk==2.20.0
setproctitle==1.3.4
Shimmy==2.0.0
six==1.17.0
smmap==5.0.2
stable_baselines3==2.4.1
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
torch==2.4.1
triton==3.0.0
typing_extensions==4.15.0
typeguard==4.4.4
tzdata==2025.1
urllib3==2.2.3
wandb==0.19.4
Werkzeug==3.0.6
zipp==3.20.2
